+++ modules/desktop/hyprland/_settings/plugins.nix
@@ -7,13 +7,16 @@ _: {
 }:
 lib.mkIf (osConfig != null) (
   let
-    inherit (pkgs.hyprworld) hyprsplitlua hypr-dynamic-cursors;
+    inherit (pkgs.hyprworld) hyprsplitlua hypr-dynamic-cursors Hyprspace;
 
     lua = import ./_lib.nix lib;
     inherit (lua) inline;
 
+    inherit (osConfig.gui) fancy;
+    inherit (osConfig.lib.stylix.colors) base03 base0D;
+
     cursorMode =
-      if osConfig.gui.fancy
+      if fancy
       then "tilt"
       else "none";
   in {
@@ -24,9 +27,9 @@ lib.mkIf (osConfig != null) (
         _var = inline ''(function() local hs = require("hyprsplit"); hs.config({ num_workspaces = 9 }); return hs end)()'';
       };
 
-      # Cursor Effects
-      plugins = [hypr-dynamic-cursors];
+      plugins = [hypr-dynamic-cursors Hyprspace];
       extraConfig = ''
+        -- Cursor Effects
         if hl.plugin.dynamic_cursors then
           hl.config({ plugin = { dynamic_cursors = {
             enabled = true,
@@ -34,6 +37,25 @@ lib.mkIf (osConfig != null) (
             hyprcursor = { enabled = true, nearest = 1 },
             shake = { enabled = true, effects = false, ipc = false },
             tilt = { activation = "negative_quadratic" },
+          } } })
+        end
+
+        -- Workspace Overview
+        if hl.plugin.Hyprspace then
+          hl.config({ plugin = { hyprspace = {
+            auto_drag = true,
+            drag_alpha = 0.4,
+            exit_on_click = true,
+            center_aligned = true,
+            hide_top_layers = true,
+            hide_overlay_layers = false,
+            show_new_workspace = false,
+            show_empty_workspace = true,
+            panel_border_width = 0,
+            workspace_margin = 10,
+            panel_color = "rgb(${base03})",
+            workspace_active_border = "rgb(${base0D})",
+            swipe_fingers = 4,
           } } })
         end
       '';