+++ packages/_module.nix
@@ -28,7 +28,7 @@ in {
           overlays =
             (attrValues config.flake.overlays or {})
             ++ [
-              (_: _: {
+              (final: _: {
                 custom = config.flake.packages."${system}" or {};
                 stagit-fork = stagit.packages."${system}".default;
                 unstable = import unstable {
@@ -39,8 +39,22 @@ in {
                 wine = windows.packages."${system}";
                 winelib = windows.lib."${system}";
                 spicetify = spicetify.legacyPackages."${system}";
-                hyprworld =
-                  hyprland.packages."${system}" // hyprsplit.packages."${system}" // hyprcursors.packages."${system}";
+                hyprworld = let
+                  hyprlandPkg = hyprland.packages."${system}".hyprland;
+                  Hyprspace = final.gcc14Stdenv.mkDerivation {
+                    pname = "Hyprspace";
+                    src = hyprspace;
+                    inherit (hyprlandPkg) version nativeBuildInputs;
+                    buildInputs = [hyprlandPkg] ++ hyprlandPkg.buildInputs;
+                    dontUseCmakeConfigure = true;
+                    installFlags = ["PREFIX=$(out)"];
+                    postInstall = "mv $out/lib/Hyprspace.so $out/lib/libHyprspace.so";
+                  };
+                in
+                  hyprland.packages."${system}"
+                  // hyprsplit.packages."${system}"
+                  // hyprcursors.packages."${system}"
+                  // {inherit Hyprspace;};
               })
               minecraft.overlay
               vscode.overlays.default